* 'Develop/Main' of http://gitlab.appinvent.in/aertrip/iOS: (46 commits)
  V1.1 Build 6
  *Offline Payment > Error state > Empty fields label/placeholder should turn red in color along with toast message
  country pointers done
  * Linked Accounts > Apple logo is missing after linking account — Done *Form > Where Search > Bottom margin at the end of the list should be 35px  ———— Done
  Fare breakup not displaying in bottom info view
  Converted keys inside rfdPlcy to double from Int Applying airline filters in multi-city causing crash
  details in Taxes, and orgPrice in breakup
  value inside Taxes struct updated to Double from int
  farePr changed to double from int
  * Form > Calendar > When user selects only check-in date and clicks on done, the check-out date should not be empty. It should be day after check-in date
  bookings date pointer done
  * Maps > Favourite Hotels > Email send option is disabled even after entering valid email * Form > Rooms Modal > Selecting more that 6 guests should only show the error. Don't select/deselect pax
  code update
  Preferences > When preferences are changed and saved show toast 'Preferences saved successfully' Preferences > When sort order is changed to last,first, for some last-name is not highlighted Preferences > Group names sorting order should be 'FirstName' and then 'A-Z' Red Error state missing - Profile.
  * Favourite Hotels > Cities sorting list is not constant. Change every time we go back and come back to Fav hotels - Done - Sorted cities by name and hotels by preferenceId
  * Edit Profile > Email & Mobile Number Field > No Validation
  1 Order should be Home, Work, Other 2 Should auto generate next tag as work after home and other after work
  no message
  Not letting save. Incorrect validation message getting generated. If blank, should delete that field and proceed
  *Filter > Date Range: Min and max Booking Dates incorrect. Min date should be first booking and max date should be last booking  in progress
  ...

+++ AERTRIP/Modules/Flight/FilghtResults/IntMCAndReturnVC/IntMCAndReturnVC.swift
@@ -333,7 +333,7 @@ extension IntMCAndReturnVC {
 
 extension IntMCAndReturnVC : flightDetailsPinFlightDelegate, UpdateRefundStatusDelegate{
    
-    func updateRefundStatus(for fk: String, rfd: Int, rsc: Int) {
+    func updateRefundStatus(for fk: String, rfd: Double, rsc: Double) {
         self.viewModel.updateRefundStatusInJourneys(fk: fk, rfd: rfd, rsc: rsc)
         self.resultsTableView.reloadData()
