+++ AERTRIP/Modules/Flight/FilghtResults/SearchResult/ResultsDataStructure/Structures.swift
@@ -45,8 +45,8 @@ public struct cityWiseAirportCode : Codable {
 
 
 public struct priceWS : Codable , Equatable {
-    var minPrice : Int
-    var maxPrice : Int
+    var minPrice : Double
+    var maxPrice : Double
     
     
     public static func == (lhs : priceWS , rhs : priceWS ) -> Bool {
@@ -283,7 +283,7 @@ struct Taxes:Codable {
 
 struct TaxesSubStruct:Codable {
     let name:String
-    var value:Int
+    var value:Double
     
     init() {
         name = ""
@@ -293,9 +293,9 @@ struct TaxesSubStruct:Codable {
 
 struct TotalPayabelSubStruct:Codable {
     let name:String
-    var value:Int
+    var value:Double
     
-    var details : [String:Int]
+    var details : [String:Double]
     
     init() {
         name = ""
@@ -376,7 +376,7 @@ struct baggageStruct:Codable  , Equatable {
 
 
 struct cancellationDetailsStruct:Codable {
-    let RAF :  [String:[String:Int]]
+    let RAF :  [String:[String:Double]]
     let SPCFEE : [String:[String:[cancellationSlabStruct]]]
     let SUCFEE : [String:[String:[sucfeeValueStruct]]]
     
@@ -404,16 +404,16 @@ struct cancellationDetailsStruct:Codable {
 struct cancellationSlabStruct:Codable{
     let slab : Int?
     let sla : Int?
-    let value : Int?
+    let value : Double?
 }
 
 struct sucfeeValueStruct:Codable{
-    let value:Int?
+    let value:Double?
 }
 
 struct refundPolicyStruct:Codable {
-    var rfd:[String:Int]
-    var rsc:[String:Int]
+    var rfd:[String:Double]
+    var rsc:[String:Double]
     var cp : cancellationDetailsStruct
     var rscp : reschedulingChargesDetailsStruct
     
@@ -437,8 +437,8 @@ struct updatedFareInfoStruct:Codable{
 }
 
 struct updatedFareInfoDataStruct:Codable{
-    let rfd : Int
-    let rsc : Int
+    let rfd : Double
+    let rsc : Double
     let cp : cancellationChargesStruct
     let rscp : reschedulingChargesStruct
